update page view for procedure and task
