+++ react/src/Components/ProceduresAndTasks/Content/index.js
@@ -1,0 +1,7 @@
+import ViewProcedure from './ViewProcedure';
+import ViewTask from './ViewTask';
+
+export {
+    ViewProcedure,
+    ViewTask
+}
