+++ react/src/Components/ProceduresAndTasks/ProcedureDialogs/ViewProcedureDialog.js
@@ -1,16 +1,12 @@
 import { amber } from '@mui/material/colors';
 import Grid from '@mui/material/Grid';
-import List from '@mui/material/List';
 import ListItem from '@mui/material/ListItem';
-import ListItemText from '@mui/material/ListItemText';
-import moment from 'moment';
 import PropTypes from 'prop-types';
 import React from 'react';
-import TextField from '@mui/material/TextField';
 import Typography from '@mui/material/Typography';
 
-import { DATE_TIME_MIN_STR } from '../../../Constants/DateStrings';
 import { SimpleModal } from '../../Utils';
+import { ViewProcedure } from '../Content';
 
 
 const ViewProcedureDialog = ({button, open, onClose, procedure}) => (
@@ -39,56 +35,7 @@ const ViewProcedureDialog = ({button, open, onClose, procedure}) => (
                 </Grid>
             </Grid>
             <Grid item xs={12}>
-                <List disablePadding dense>
-                    <ListItem>
-                        <ListItemText
-                            primary="Created at"
-                            secondary={moment(procedure.created_at * 1000).format(DATE_TIME_MIN_STR)}
-                        />
-                    </ListItem>
-                    <ListItem>
-                        <ListItemText
-                            primary="Documentation"
-                            secondary={procedure.doc || '-'}
-                        />
-                    </ListItem>
-                    <ListItem>
-                        <ListItemText
-                            primary="Procedure arguments"
-                            secondary={`[${procedure.arguments}]`}
-                        />
-                    </ListItem>
-                    <ListItem>
-                        <ListItemText
-                            primary="Definition"
-                            secondary={procedure.definition ?
-                                <TextField
-                                    fullWidth
-                                    multiline
-                                    name="procedure"
-                                    type="text"
-                                    value={procedure.definition || '-'}
-                                    variant="standard"
-                                    InputProps={{
-                                        readOnly: true,
-                                        disableUnderline: true,
-                                    }}
-                                    inputProps={{
-                                        style: {
-                                            fontFamily: 'monospace',
-                                        },
-                                    }}
-                                    InputLabelProps={{
-                                        shrink: true,
-                                    }}
-                                />
-                                : '-'}
-                            secondaryTypographyProps={{
-                                component: 'div'
-                            }}
-                        />
-                    </ListItem>
-                </List>
+                <ViewProcedure procedure={procedure} />
             </Grid>
         </Grid>
     </SimpleModal>
